+++ example/lib/navigation_rail_dropin_inverted.dart
@@ -1,0 +1,77 @@
+import 'package:flutter/material.dart';
+import 'package:curved_navigation_rail/curved_navigation_rail.dart';
+
+/// Drop-in replacement for NavigationRail using the official example at
+/// https://api.flutter.dev/flutter/material/NavigationRail-class.html
+///
+/// ** NOTE ** labelType parameter is currently not supported.
+///
+/// This is the inverted version which is for right to left directionalities.
+/// TODO: Add directionality support and update this example.
+void main() => runApp(const MyApp());
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  static const String _title = 'Flutter Code Sample';
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      title: _title,
+      home: MyStatefulWidget(),
+    );
+  }
+}
+
+class MyStatefulWidget extends StatefulWidget {
+  const MyStatefulWidget({Key? key}) : super(key: key);
+
+  @override
+  State<MyStatefulWidget> createState() => _MyStatefulWidgetState();
+}
+
+class _MyStatefulWidgetState extends State<MyStatefulWidget> {
+  int _selectedIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Row(
+        children: <Widget>[
+          CurvedNavigationRail(
+            selectedIndex: _selectedIndex,
+            onDestinationSelected: (int index) {
+              setState(() {
+                _selectedIndex = index;
+              });
+            },
+            destinations: const <NavigationRailDestination>[
+              NavigationRailDestination(
+                icon: Icon(Icons.favorite_border),
+                selectedIcon: Icon(Icons.favorite),
+                label: Text('First'),
+              ),
+              NavigationRailDestination(
+                icon: Icon(Icons.bookmark_border),
+                selectedIcon: Icon(Icons.book),
+                label: Text('Second'),
+              ),
+              NavigationRailDestination(
+                icon: Icon(Icons.star_border),
+                selectedIcon: Icon(Icons.star),
+                label: Text('Third'),
+              ),
+            ],
+          ),
+          // This is the main content.
+          Expanded(
+            child: Center(
+              child: Text('selectedIndex: $_selectedIndex'),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}