+++ test/curved_navigation_rail_test.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:curved_navigation_rail/curved_navigation_rail.dart';
+
+void main() {
+  test('', () {});
+}